refactor: context checking state as derived store Signed-off-by: Philippe Martin <phmartin@redhat.com>
