+++ packages/renderer/src/lib/preferences/PreferencesKubernetesContextsRendering.svelte
@@ -4,7 +4,7 @@ import { Button, EmptyScreen, ErrorMessage, Spinner } from '@podman-desktop/ui-s
 import { onMount } from 'svelte';
 import { router } from 'tinro';
 
-import { kubernetesContextsCheckingState, kubernetesContextsState } from '/@/stores/kubernetes-contexts-state';
+import { kubernetesContextsCheckingStateDelayed, kubernetesContextsState } from '/@/stores/kubernetes-contexts-state';
 
 import { kubernetesContexts } from '../../stores/kubernetes-contexts';
 import { clearKubeUIContextErrors, setKubeUIContextError } from '../kube/KubeContextUI';
@@ -14,8 +14,6 @@ import SettingsPage from './SettingsPage.svelte';
 
 $: currentContextName = $kubernetesContexts.find(c => c.currentContext)?.name;
 
-let previousState = new Map<string, 'waiting' | 'checking' | 'gaveup'>();
-let checkingCount = new Map<string, number>();
 let kubeconfigFilePath: string = '';
 
 onMount(async () => {
@@ -30,25 +28,6 @@ onMount(async () => {
     kubeconfigFilePath = 'Default is usually ~/.kube/config';
   }
 });
-
-$: {
-  for (const [context, state] of $kubernetesContextsCheckingState) {
-    if (!previousState.has(context) || previousState.get(context) !== state.state) {
-      if (state.state === 'checking') {
-        const prev = checkingCount.get(context) ?? 0;
-        checkingCount.set(context, prev + 1);
-        checkingCount = new Map(checkingCount);
-      } else if (state.state === 'waiting' && (checkingCount.get(context) ?? 0) > 0) {
-        const prev = checkingCount.get(context) ?? 0;
-        setTimeout(() => {
-          checkingCount.set(context, prev - 1);
-          checkingCount = new Map(checkingCount);
-        }, 2000);
-      }
-      previousState.set(context, state.state);
-    }
-  }
-}
 
 async function handleSetContext(contextName: string) {
   $kubernetesContexts = clearKubeUIContextErrors($kubernetesContexts);
@@ -181,7 +160,7 @@ async function handleDeleteContext(contextName: string) {
                       UNKNOWN
                     {/if}
                   </div>
-                  {#if (checkingCount.get(context.name) ?? 0) > 0}
+                  {#if $kubernetesContextsCheckingStateDelayed?.get(context.name)}
                     <div class="ml-1"><Spinner size="12px"></Spinner></div>
                   {/if}
                 </div>
